fix float to int

+++ python_introduction/finance_calculator.py
@@ -7,6 +7,6 @@ monExpenses = int(monExpenses)
 interest = .05  # Example interest rate of 5%
 monSavings = monIncome - monExpenses
 
-projectedSavings = monSavings * 12 + (monSavings * interest * 12)
+projectedSavings = int(monSavings * 12 + (monSavings * interest * 12))
 print(f"Your monthly savings are: ${monSavings}.")
 print(f"Projected savings after one year, with interest, is: ${projectedSavings}.")
